doCollectionEdit: add ability to skip sanitization
I don't the sanitization code is correct, because if the url was correct to begin with, but we happen to not resolve yet (e.g. api down), we ended up with an edit list that is full of nulls.

Not sure what the full fix is, so just constrain this to Featured Channels for now, because there is a higher chance of not resolving the list because we skip the Publish Form.