Fix thumbnail fetch for reposted image (without `thumbnail.url`)
## Issue
If an image file does not contain `thumbnail.url`, the fallback is to fetch the image itself as the thumbnail. But for the case of a reposted image file, the `get` API ends up throwing a "stream doesn't have source data" error because the report uri was used.

Also, `claimPreviewTile` was missing the `streamingUrl` selector like it's `claimPreview` counterpart.

## Fix
1. Use the repost src uri when doing `get`.
2. `claimPreviewTile`: get `streamingUrl` prop.
3. Change the `streamingUrl` prop to look at the repost src instead.