Tweak confusing "use browser default" behavior
## Issue
Consider this scenario:

1. Set browser language to:
    - Firefox = Japanese
    - Chrome = English
2. Restart both browsers.
2. In both browsers (or let them sync naturally), set language Spanish.
3. In Firefox, change the language to Japanese and save.
    - Chrome becomes English when sync completes.

The code sets the language to `null` when it matches the browser's language.
But I don't think the user will understand the behavior since they specifically se to Japanese.

## Attempt 1
My initial solution involved adding a "--- Use browser default ---" selection into the dropdown to explicitly set the setting to `null`, instead of implicitly checking against browser language.

It works, but it is ugly and involved too many places. It's also error-prone from the coding standpoint, because we now need to query 2 selectors instead of 1.

## Attempt 2
Just don't set to `null` once the user have changed a language one time. "Setting to X" now really means "Setting to X", instead of "potentially setting to browser language".

Other apps seem to do the same.