Fix language toggling back-and-forth - reason 1 (cross setting)
## Reproduce
1. Have 2 tabs opened -- Settings Page, Home Page.
2. Change the language in Tab_1.
    - The dropdown will auto-toggle 1 more time before settling down.

## Reason
- This useEffect is meant to only fetch the language-json when the setting is directly changed without calling doSetLanguage, say from a sync action.
- But it incorrectly called doSetLanguage, so the tabs end up forcing each other to sync, and the value somehow briefly reverted to the wallet's value.

## Fix
Changed to doFetchLanguage as it was intended.

## Future Fix
doSetLanguage is problematic because not all parts of the code can call this, and clients shouldn't need to know about the internal workings of having to fetch the json.

A safer way to change language is to invert the sequence:
1. Let `doSetClientSetting(SETTINGS.LANGUAGE, lang)` be the only way to change the language setting.
2. Then, add a redux middleware to perform the language-json-fetch behind the scenes when SETTINGS.LANGUAGE changes. This is the "redux way", I believe.

But this is pretty much a re-do, so holding off until more corner-cases arise (probably no more).