Extend customized-homepage setting to Sidebar
Ticket: 1203

## Approach
For now, make the customize-homepage setting the single source of truth, instead of duplicating 2 sets of settings.

Sidebar: needed a place to put the "customize" button, so I had to introduce a section header to house it. The "manage" button for subscriptions is now styled the same way, which I think looks more structured than having it at the bottom.

Note: "Following" and "FYP" will not appear in the section, because the former is locked at a higher section (honoring that for now) and the latter doesn't have a dedicated page at the moment. Users should be able to figure out what's happening.

Note: When all categories are removed, I left the header there so it's easier to put categories back in. Could be removed if needed, since it's easy to access the customize button from the homepage.