Fix language toggling back-and-forth - reason 2 (sync-lock)
## Reason
Sync-lock broken due to global re-render from a language change.

## Issue
Syncing should be disabled in the Settings Page, but was indirectly re-enabled due to the re-mounting of Settings Page that happened when the language changed.

## Change
In general, the topmost App component renders everything when the language changes. This avoids the need for _every_ component to listen to `selectLanguage`.

But we don't want Settings Page to be part of this, because the unmounting causes the sync-lock to be opened.

Addressed this by making Settings Page an exception that uses the old method of listening to `selectLanguage` directly for language updates.