Re-do: fix volume slider issues
## Original Issue
Dragging the volume slider in Floating mode will drag the player as well.

## Initial Fix
18debf51

## Current Issue
- https://sentry.lbry.tech/organizations/odysee/issues/22695
- 1735 - The volume hover seems to not load properly until clicked

While the override idea came from the videojs folks, the new object isn't completely populated, the old object seems to still linger (we didn't dispose), and there seems to be some internal linkage to the old object (not sure how to sever).

## Alternative
From the `Draggable` component, we can examine where the event propagated from, so instead of stopping the propagation from videojs, we just absorb it here and cancel off the Floating Player drag action.