SignOut: prevent infinite loop from 'beforeUnload' listener
## Ticket
419

## Issue
We use 'beforeUnload' to invoke the browser's alert message when there are pending actions like Uploading or unsaved Settings changes.

The sign out code have no idea this is going on, so when it tries to reload, the alert appears in a recursion.

## Approach
- Add a wrapper for the 'beforeUnload' functionality, so we can keep track of it.
- Before signout out, open a modal and list out all the pending actions.
    - This also solves the problem of the browser alert not saying anything useful to the user (all browsers no longer relay the msg param and just show a generic one to avoid spam abuse).
- If OK is pressed, we unregister the listeners and proceed as normal.

## Caveats
It would be nice if we could show the same modal on an F5 case, but I don't know how to intercept that case. Maybe in the future.