TagSearch: remove channel limit and release time
Went with the opt-in hook method rather than applying it globally in ClaimListDiscover -- it's easier to control, change and test.

For example, we can't apply the removal by simply checking if a tag exists or not, because a livestream ClaimListDiscover includes an internal tag. We could add code to ignore internal tags, but that's starting to put too much logic into ClaimListDiscover (not flexible), hence the hook method.

We could also tweak further, like maybe just increasing the channel limit rather than deleting ... sometimes, results still overwhelmed by some channels.

It's currently applied to:
- Category Page (minus the livestream section)
- Following Page
- Channel Page