Category: Fix "Top + This Week" looking until 1995
Patch for c4f88ac7

## Root-causes
(1) `getReleaseTime` not able to detect Order=Top reliably due to implicit defaults that's not always reflected back to URL params.
(2) `getReleaseTime` not able to determine if "This week" was a user value, or auto-defaulted value.
    - But it didn't have to. Just skip and use base logic when Order=Top.

## Change
Make Discover define what the default order is, so the logic in `getReleaseTime` doesn't have to guess.