TagSearch: open up Freshness if Top is selected
## Issue
We previously removed the channel limit and release time when there is tag query. This causes the Freshness drop-down to be non functional.

## Change
Instead of graying it out, make it functional if Top is selected.

Rationale:
Right now, the Order switches to non-Top automatically when the tag query is entered (legacy behavior). `release_time` will still be removed for these cases. But if the user explicitly changes to Top, seems logical to give user full Freshness filtering options since the drop-down is already there.