Publish:  Check `status` indefinitely
Rather than stopping after N tries and letting the user manually check afterwards, just keep on checking as long as the status is 202 (pending/alive).

The halt was displayed through an error dialog, which confused many users.

On F5, pressing Check Status will restart the indefinite check.

## Approach
Since negative number is truthy in Javascript, just send "-1" as the retry count and let it deduct indefinitely. If we want to set a limit again in the future, then it would just be a variable change.

Theoretically, it would overflow, loop back and hit 0, but 9007199254740991 retries should be a lot...