Alleviate emoji img loading
## Issue
All emoji image files are being fetched on F5

## Root
- Something is causing `loading="lazy"` to not work reliably. I'm guessing it's due to ReachUI's Tab.
- Also, EmojisPanel and StickersPanel currently requires all emojis to be mounted for the "scroll to category" to work correctly.

## Approach
Until those components support pagination, alleviate the problem by delaying the mounting of these components for the first time.

While all the images are still being fetched when the panel is opened, at least it is not loaded on F5 where most people aren't commenting directly.

One downside to the approach is that there will be a lag when opening the panel for the very first time. Subsequent ones would be fine.