Clarify concatenated string to the translator
Without creating another contextual string, this is the simplest approach to tell the translator what the front string is going to be.

The 2 strings can be concatenated seamlessly in English, but I don't think it will work for all languages. Since translation activity has died down at the moment, we'll handle it later if/when a translator complains.