Fix selector violation in portal fallback
## Issue
Selectors must not mutate the data source. This violates the natural data flow, and it doesn't make sense in general for a getter function to alter the storage.

## Change
Normally, one would clone, derive and cache (if needed) the data in the selector. But since we want to fallback to always happen in our app (i.e. no cases where we need the original field to be blank), we'll post-process the fetched data directly so we don't have to do it at the selector side.