Support language specific `/$/content/get` (#2575)
* doFetchHomepages: update comments

Almost forgot why we are still fetching when we have that flag turned on.

* Encapsulate homepage storage

This encapsulates and minimizes exposure of the fetched homepage data to avoid accidental changes (see next commit). The GUI should be accessing data from selectors anyway, and this also makes it easier to change storage location later.

* Fix selector violation in portal fallback

## Issue
Selectors must not mutate the data source. This violates the natural data flow, and it doesn't make sense in general for a getter function to alter the storage.

## Change
Normally, one would clone, derive and cache (if needed) the data in the selector. But since we want to fallback to always happen in our app (i.e. no cases where we need the original field to be blank), we'll post-process the fetched data directly so we don't have to do it at the selector side.

* Content API: add "?hp=xx" for specific homepage query

2388

## API Notes
To retrieve full homepage dataset:
    - https://odysee.com/$/api/content/v2/get

To retrieve a specific homepage:
    - https://odysee.com/$/api/content/v2/get?hp=xx (where xx = en|fr|es|...)
    - The return object will still include the keys for all available homepages, but the values for unwanted homepages will be set to `null`. The full key set will allow clients to generate drop-down lists without needing a different API.

* Utilize "?hp=xx" in Odysee

Previously, doFetchHomepages() performs 2 things:
- load from built-in
- fetch and update with latest one (in case the built-in is from browser cache)

Now, split this up to 2 separate functions:
- doLoadBuiltInHomepageData: called during startup
- doFetchHomepages: called at least once per startup, and on subsequent homepage-selection changes.