Fix hideSort attribute ended up in the DOM
## Issue
The newly-introduced `hideSort` parameter ended up in the DOM

## Change
This is the problem with `getSidebarCategories` partially grabbing values from the Category Metadata. I think we can inverse the logic to grabbing specific items that we need instead (e.g. `link`, `title`, `route`, `icon`, etc.), but too lazy to re-test everything, so just continue with the existing filter pattern for now.