Upgrade video.js 7.21.4; install resolution to lock version
Otherwise, other plugins will try to pick version 8 if the plugin supports that. This causes 2 incompatible version of video.js to be installed.

The recent reset of yarn.lock caused this. The solution is to add a `resolution`, and that should be the flag to change when upgrading in the future.