Don't mount video until `get` is complete
## Ticket
2624

We were seeing video urls with `undefined` being called, before the actual url is called later.

## Cause
PR_2578 changed the "loading" if-block such that there is no final else-block, so the bottom code got executed and mounted the video player.

## Change
Based on the original code, it was just showing the spinner whenever there is no `fileInfo`, so I followed that logic.