iOS: skip autoplay countdown for now
This is to circumvent the problem of Safari blocking non-muted autoplays without user interaction. Seems like the blocker will also activate if there is an async action between the click handler and play(), or some major changes happened to the player configuration.

The minimal-change workaround is to simply not show the countdown for now, until we have a way to better orchestrate the setup sequence.

Reference:
https://arrangeactassert.com/posts/how-to-fix-the-request-is-not-allowed-by-the-user-agent-or-the-platform-in-the-current-context-possibly-because-the-user-denied-permission/