Handle safari-mac chapters
## Ticket
2638

## Change
I believe the issue is with `addTextTrack` not having cleanup counterparts + we are now recycling the video instance + the Safari implementation probably didn't handle it robustly.

Moved to `addRemoteTextTrack`, which the docs says it includes auto cleanup when the src changes. It is also the recommended API anyway as it comes with `removedRemoveTextTrack`, in case we ever need to manually clean things up.

But while doing so prevented the crash, the player stops working on autoplay next. Disabling chapters on Safari until we figure out what other "manual cleanup" is required for this damn single-instance implementation.