Fix abandoned claims showing up in Recommended
## Ticket
2607

## Change
Undo a line introduced in e1b13013 that's causing the problem.

Logically, `shouldHide` can be because of `!claim`, so adding a `claim` there doesn't make sense.

But it's unclear what the change was trying to fix, so this could re-open some issues. But whatever that issue is, it'll require a different fix.