Add TimeMarker component to handle clipping chapter titles
## Issues
1. With the use of `clip-path` to get the island effect on the progress bar, the time/chapter tooltip gets clipped as well.
2. There is an existing bug when time/chapter tooltip is clipped off screen on the extremes.

## Change
To address the first problem, we need to move the component up in the z-order. Creating a custom component is the only way, since the defaults rely on a very specific hierarchy to work. Fortunately, the component is simple enough to clone.

Since we are already cloning, we can also address the screen clipping problem (we have full control of the component).