Chapter markers: replace dots with island effect
I still think that we should not touch the existing ProgressControl (which handles a complicated set of subcomponents) until we have a real need to customize the progress bar entirely.

This solution is a css-only one, where we add a `clip-path` to SeekBar to produce the island effect. This way, the videojs side remains untouched.