Clean up Publish-slice types
It's a pain having to maintain both PublishState and UpdatePublishFormData. The latter should essential just be the former but in optional form, so use flow's `$Shape` to automatically generate that.

`$Shape` is deemed deprecated. Tried to upgrade flow, but it involves a lot of fixes to make flow's new "types first" paradigm work. Since the time would be better spent converting to Typescript, just use `$Shape` for now.

Marked a few new warnings as todos -- will address them in the next tasks. Don't want to include any logic changes for now (limit to just flow changes) so we don't have to re-test anything.