Fix subscription claim_id parsing
## Ticket
2659 - Channel names with "*" causes issues.

## Root
```
export function splitBySeparator(uri: string) {
  const protocolLength = 7;
  return uri.startsWith('lbry://') ? uri.slice(protocolLength).split(/[#:*]/) : uri.split(/#:\*\$/);
}
```

This function splits * (on top of # and :), and was being used to parse subscription uris. I think * is valid in channel names?

## Related issue
`parseUri` is expensive, and we constantly re-parse the subscription uris to get the IDs.

## Change
Use `parseUri` instead of `splitBySeparator`, but cache the results through a selector.