doUpdateCreatorSettings: alter 29f7c7af
Doing a try-catch block here will simply cancel off the effort to catch and relay the error to the user.

Based on 29f7c7af's commit message, it's was trying to fix "TypeError: load failed". If that is the case, the fix doesn't make sense because neither `dispatch` nor `doToast` will throw that error -- it's usually a `fetch` error in iOS.

If that block is indeed throwing an error, it's most likely just because `err` is null (someone reject with null?), so changed the code to handle that.

I believe the true cause for "TypeError: load failed" is the recent change in doFetchCreatorSettings that re-throws the Error in the catch block. Let's handle that in the next commit...