doFetchCreatorSettings: fix unhandled catch block
## Issue
Lots of Sentry errors from this unhandled error, which is usually just due to temp network glitch.

The root-cause is that `doFetchCreatorSettings` was changed to re-throw errors so that a shortcut `then/catch` can be done directly at the React component. Anti-pattern arguments aside, none of the callers handled the `catch` block, and in fact relied of the unhandled error as a means to stop the comment from being sent.

## Change
Always handle `catch` when there is a `then`. Display proper toast.