injectPinUrls: fix invalid change in fbe5eb38
## Original issue
I believe the original issue was that "an empty category but with pins" will end up producing an empty list instead of a combined one. The original code short-circuited on empty arrays.

## The previous fix
I don't think it would have worked. It is mixing mutable vs. immutable change here. Also, returning from the function does nothing because the caller is expecting a mutable change.

## Change
Re-do the fix per original intention of the function. Removing the short-circuiting is all that is needed.