Lint-rule changes
Make these as 'error' instead of 'warning', else no point.

-- space-before-function-paren --
since majority of the code, and prettier (?) is formatting anonymous functions with a space, set the rule as "always" to sync up.

-- react-hooks/exhaustive-deps --
These errors are almost always valid, so it should be an 'error'.

Unfortunately, it's either all or nothing in terms of annotating. We can't leave the dep-array blank because we can't tell whether that's intentional or forgot-to-update.