Improve buildURI assertions
The typical usage of buildURI is to wrap it inside a try-catch or check the results with isURIValid.

New users typically don't know this, but previously we just warn using a console message.

After we moved to asserts, we have to disable asserts for those correct usages (already checking output), but continue to assert on the rest.