Save validated access keys for auto-append later
If we navigated away from an unlisted uri, and return (say, from the Floating Player), the access key in the URL will be lost since navigation is based on claim uri.

Instead of changing now the navigation system works, we stash the validated keys the first time it was accessed. The GUI side that does the validation when navigating can check if a cached key exists. This also saves the need to call channel_sign again.

This commit only covers the caching part. The GUI part comes next...

----

I also introduce selectContentStates() to have direct access to states. As noted in the Redux documentation, selectors are not only needed if there is transformation. Direct-access selectors should eventually be removed.