Add general-purpose ClaimSearchView
View for claim_search results, encapsulating the pagination and API calls.

It's a simplification of ClaimTileDiscover/ClaimListDiscover that has gotten bloated with lots of confusing props.

## Misc
Also stashed the page information from the `claim_search` result. I think that gives a better indication of whether we are at the final page, vs. trying to figure out through the modulus calculation.