Flow: fix react-beautiful-dnd
## flow-typed/npm/react-beautiful-dnd
Add stub module so we don't have to add Flow suppressors when importing.

## flowconfig
- Don't ignore the package so that types can be automatically inferred.
- Ignore 'raf-schd' instead (packaged used by dnd), since it's one level down. Shouldn't have happened, but not sure what's going on.

## sortableList, homepageSort
Addressed the Flow warning on 'style' potentially not the same type as what the code expects.

The useRef usage is wrong (should use ref.current), but Firefox seems to rely on the incorrect code. Could set `left = undefined` to make it look "correct", but leaving things in original form.

## claimList
Similarly handled 'style' like previous case. But for the bottom part of the code, I don't want to re-test all clients, so just suppressed the warning and added an assert to keep this a "flow only" change.