iOS: Enable quality selector with "Original" and "Auto" only
## Ticket
2705

## Change
If the video is HLS and the platform is iOS, make the quality selector display "Auto" (for native) and "Original".

This is to handle the poor performance of the native algorithm which often selects the worst quality. Users can now at least try the original resolution.

Ended up choosing 'loadedmetadata' for the action, since we also need to make the `swapSrcTo` call. This event is later than SRC_CHANGED, and feels safer.