Fix doClaimSearch signature
## Issue
`settings` shouldn't have a positive default value like `fetchStripeTransactions: true`, because the parameter is supposed to be a incremental addition of features.

For example, if I provide `settings = { autoPaginate: true }`, this indirectly makes `fetchStripeTransactions = false`, which doesn't make sense.

## Fix
Since the current behavior is to fetch purchases ubiquitously, removed the optional setting for now. If there is a need, we can add a `noStripeFetch` in the future.

This should be a safe change since it also checks if the claim has any fiat tags before doing the additional fetch.