Replace useFetchViewCount with inline fetch
## Issue
useFetchViewCount was meant to be a local (component) cache of uris that needs to have its views fetched. But it brought in `claimsByUri` which gets invalidated a lot, causing the component to get updated unnecessarily.

It doesn't need to bring in `claimsByUri` actually, as the Action side has full access to the store to perform the uri-to-id translation

## Completely different approach
But we can avoid the need for the hook entirely by making `doClaimSearch` dispatch the call. It knows exactly what are the new IDs, so we can also avoid the need to filter and cache.