useResolvePins: don't bring in claimsById to avoid over-renders
## Issue
`claimsById` will be invalidate on any `claim_search`, so bringing this in causes the parent to update too much.

## Change
Grab the results from `doResolveClaimIds` directly from the `then` block.

## Dirty notice
This is arguably anti-pattern because we are reaching into Redux in the middle of an action, but it's better to be pragmatic in this case. We are just interested to know if the results exists for this uri, so we aren't affected by any transformations that happen in the reducer.