Handle crash when the cost info cannot be determined
## Issue
LBRY-DESKTOP-WEB-XKY

Not sure what's the actual case, but one way to reproduce is by blocking the `exchange_rate` call.

## Change
Just show a '?' in this case. That should make the user refresh, and even if they didn't, the backend API will produce an appropriate error toast. I don't think we need to change the GUI for this rare corner case.