Set locale only once globally per language change
Components can skip listening to language state if the value is just needed to determine locale. We can set it one time during the change, and the entire app is already re-rendering on language changes.

-- Aside --
"Chrono" is picked the name for a future wrapper for any date-time management.

moment will most likely need to be replaced soon since it is no longer maintained (potential security risk?).

Having the wrapper would make it easier to swap things out.