Persist 'uriAccessKeys' (QA-71)
## Issue
Right now, each time we do a new signature, it will generate a new link because the signature and timestamp change. In the future this may change.

## Change
The signature is alreayd being stored in redux, so as long as the service returns the same short URL given the same payload, we are good. Just needed to make the redux state persistent.

The service currently does seems to return the same URL, so it does implement `findIfExists` behind the scenes.