comment.List - stop updating channel settings
## Issue
- The code is in a weird state ... the action wants to update the settings, but reducer doesn't update with new data correctly.
- The update (albeit incorrect) was always executed even if the value is the same, causing the UI to update unnecessarily.

## Change
The original intention for this piece of code is to update `comments_enabled` directly if `comment.List` says it is disabled, instead having to waste another `settings.List` call.

The membership change broke this, and later someone else seem to have added the `settings.List` call in CommentCreate, so it's pointless to retain it anymore.