Fix copyright always reset to "All rights reserved"
09c0315e is doing that, but no reason given in the commit message. Guesses:

## Wanted 'Copyright' and 'Other' to have different strings?
If that is the create, create another state instead of sharing it.

Looking at the rest of the code, it doesn't seem like that is the intention, though.

We also shouldn't be purging user-entered values when drop-down changes. When they switch the dropdown back to the previous value, the related fields should come back.

## Wanted to reset the field on page entry?
If there was a need to reset the values, do that from redux (see `doPrepareEdit`, `doClearPublish` and similar actions).

--- General Note:
It's probably also a misunderstanding of the Redux data flow. For the case of Publish components, there shouldn't be logic to change states on their own -- redux side owns that logic.

Components only (1) translate state to jsx (2) relay user's changes to redux.

When both sides try to own the logic, it gets hard to maintain.