Remove redundant isMarkdownPost
The way `isMarkdownPost` was being updated from 3 components feels a bit dirty and unintuitive. The data flow became 2-way.

Since we are already using Redux as the orchestrator, plus there is also a `type` state for us to infer if it is a markdown, `isMarkdownPost` is redundant.

As long as we follow the design and use doBeginPublish/doPrepareEdit to start the forms, we don't need to update the states from the components.