Remove redundant isMarkdownPost
`isMarkdownPost` was being managed by 3 components, and not in a "Redux/Flux way".

Since we are already using Redux as the orchestrator, plus there is also a `type` state for us to infer if it is a markdown, `isMarkdownPost` is now redundant.

As long as we follow the design and use doBeginPublish/doPrepareEdit to start the forms, we don't need to update the states from the components.