Publish: fix junk 'filePath' from rehydration
## Issue
2722
After selecting a file and hitting F5, the form still thinks it has the file and allowed Submit.

Browsers require file-selection to be explicit -- can't be chosen programmatically.

## Why
Redux cannot rehydrate un-serializable objects like `File`.  When rehydrated, we get an empty object which is truthy, so the form wasn't disabled.

## Change
Ensure `filePath` is always cleared in reload.