Fix incorrect "Cannot set future date" for replay uploads
## Issue
The form starts off with "Update Only", which automatically put release_time in the future.  When switching to "Upload Replay", the calendar is not applicable here but the form is still blocked with "Cannot set future date"

## Root
Misunderstanding of what `isLivestreamPublish` meant, which does not cover "Upload Replay".

## Change
Use the right flag.