Fix livestream replay state handling
## Issue
When selecting a replay in livestream edit, but then deciding to just update the stream info instead, it will still upload the replay.

## Root
There are 3 styles of states surrounding this feature that try to govern the same thing. Most likely coming from separate developers not realizing the form states are in Redux.

## Overall Change
Restore things back to Redux as the sole owner. It needs to be that way anyway so that the Publish Action has the full picture to generate the payload. Trying to synchronize the states between 2 sides is just confusing.

## Change Notes
1. For 'LiveCreateType' and 'LiveEditType', the values are more verbose than usual to avoid repeating the issue again.

2. There were several useEffects that changes another state when one state changes. Scattering this logic around the View is bad -- hard to find or realize it exists. As Redux is the owner, do the auto-adjustment in the reducer of ACTIONS.UPDATE_PUBLISH_FORM like the rest (or split into another action if necessary).

3. For this particular form, I choose not to do auto-adjustments. e.g. When user changes a radio selection back and forth, the previous sub-values should be repopulated, so we shouldn't clear it. We just need to grab the right fields when generating the payload.

4. 'isLivestreamPublish' can now be safely removed since we can infer that from the new states. It was a poorly-named one that was used to indicate if the Livestream publish was creating a placeholder or not.

5. Finally, handle the original problem statement by only adding `remote_url` only when needed.