remoteFileUrl reset logic
Ideally, the selected `remoteFileUrl` should be persistent so that when "replay" is selected again, user don't need to re-select the replay-url from the list of radios.

But the component currently doesn't repopulate (needs to find the radio index during mount), so just clear the url for now when the mode is no longer "replay".