ScheduledStream: don't update when transitioning to a new search
## Issue
If blocklists come in late for example, the search results is `undefined` because we have a new query. But most of the time, the results are the same, causing unnecessary erasure.

## Change
Skip this update. Eventually, the results will either be [...] or null, which then a new render display the right results.

## Note
I couldn't figure out how to ensure type safety, so one must ensure new props are compared in `areEqual`