Make 'type' an automatic variable based on navigation changes
## Issue
`type` used to be set in doBeginPublish, but it gets un-synchronized when doing `POP`, F5, or direct URL access.

When un-synchronized, the SDK payload logic becomes incorrect.

## Approach
Since the "Submit" button is currently tied to the current page
(i.e. no floating Publish forms), and that all 3 forms share the same states, we need `type` to always be correct as the reference variable for the rest of the logic here.