Fix non-exclusive-content perks appearing in form
## Issue
2835
> Changing channel in form while restricted settings is expanded, allows selecting tiers with no "exclusive content" perk

## Root
Using vanilla JS in React leads to update issues like this.

But I understand why it was done that way -- the color of the tiers is tied to the css, which restricts how the elements need to be structured (i.e. `.checkbox` needs to be mounted but hidden, and next to each other). We also don't want to tweak FormField for something so specific, so it's an ok tradeoff.

## Change
Ideally, the color should be part of the membership object, not by how the HTML is structured.

But since the filtered array needs to be memo'd anyway, use that as a driver for the effect. Hopefully it's good enough to cover all cases.