Fix restrictions not cleared when switched to Unlisted
Note: yup, there is a DRY problem here, but `doPublish` and `doPublishResume` had it for a while.

Wanted to pull out the duplicate logic to a thunk, but I'm not sure how `getState` will be captured in a success callback -- is the entire current state captured or just the getter function?

Anyway, feels safer to extract the values first before creating the closure, as the user could immediately start another publish and alter the form states.