Publish: fix missing Restriction card after F5
## Issue
- 2421: "Restrict content" section gets lost on refresh or via direct link
- 2150: "Edit" still doesn't retain original channel value #

## Root
The ChannelSelector's `autoSet` mechanism relies on `channelToSet` being correct when the ChannelSelector is mounted. In the reloading case, the channels are still being fetch, so `channelToSet` was undefined. This sets the ChannelSelector to Anonymous, which then the Restriction Card hides itself.

The status doesn't correct itself because `autoSet` is only executed on-mount.

## Change
Only change the current channel when trying to Edit. The value needs to start with the claim's original channel. User can change this later.

As `claimToEdit` is stored and rehydrated on F5, we have the right value to cover the F5 case as well.