Use immer in Reactions slice.
## Why
Discussed with team and decided to use immer moving forward to solve the following:
- Easier to write immutable code (no more spreading everywhere)
- If the reducer involves conditionals and ended up not changing the value, the original reference will be returned to save up one render.

## Change
Using a non-critical functionality (file reactions) as a trial-run.

Browser-support for `Proxy` is just slightly less than `Promise.all` that we've been using, so it shouldn't impact any of our existing users, but wanted to be sure.